aws_iam_mfa_restriction/requirements.txt
-e .
attrs==19.3.0
aws-cdk.aws-iam==1.51.0
aws-cdk.cloud-assembly-schema==1.51.0
aws-cdk.core==1.51.0
aws-cdk.cx-api==1.51.0
aws-cdk.region-info==1.51.0
cattrs==1.0.0
constructs==3.0.4
jsii==1.9.0
publication==0.0.3
python-dateutil==2.8.1
six==1.15.0
typing-extensions==3.7.4.2
